[Mod] 更新到1.18.2.5.2
增加linux支持
