+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# # 1.18.2.5.2版本
+
+1. 增加linux支持
+
 # # 1.18.2.5.1版本
 
 1. 适配vnpy框架4.0版本
