+++ README.md
@@ -4,8 +4,8 @@
 </p>
 
 <p align="center">
-    <img src ="https://img.shields.io/badge/version-1.18.2.5.1-blueviolet.svg"/>
-    <img src ="https://img.shields.io/badge/platform-windows-yellow.svg"/>
+    <img src ="https://img.shields.io/badge/version-1.18.2.5.2-blueviolet.svg"/>
+    <img src ="https://img.shields.io/badge/platform-windows|linux-yellow.svg"/>
     <img src ="https://img.shields.io/badge/python-3.10|3.11|3.12|3.13-blue.svg" />
     <img src ="https://img.shields.io/github/license/vnpy/vnpy.svg?color=orange"/>
 </p>
