+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "vnpy_da"
-version = "1.18.2.5.1"
+version = "1.18.2.5.2"
 description = "DA International gateway for VeighNa quant trading framework."
 readme = "README.md"
 license = {text = "MIT"}
@@ -9,6 +9,7 @@ classifiers = [
     "Development Status :: 5 - Production/Stable",
     "License :: OSI Approved :: MIT License",
     "Operating System :: Microsoft :: Windows",
+    "Operating System :: POSIX :: Linux",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -21,8 +22,7 @@ classifiers = [
 ]
 requires-python = ">=3.10"
 dependencies = [
-    "vnpy>=4.0.0",
-    "wmi"
+    "wmi; sys_platform == 'win32'"
 ]
 keywords = ["quant", "quantitative", "investment", "trading", "algotrading"]
